requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py=0.7.1=pypi_0
astor=0.7.1=pypi_0
atari-py=0.1.7=pypi_0
ca-certificates=2019.1.23=0
certifi=2018.8.24=py35_1
chardet=3.0.4=pypi_0
cycler=0.10.0=pypi_0
future=0.17.1=pypi_0
gast=0.2.2=pypi_0
grpcio=1.20.1=pypi_0
gym=0.12.1=pypi_0
h5py=2.9.0=pypi_0
idna=2.8=pypi_0
keras-applications=1.0.7=pypi_0
keras-preprocessing=1.0.9=pypi_0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libstdcxx-ng=8.2.0=hdf63c60_1
markdown=3.1=pypi_0
matplotlib=3.0.3=pypi_0
mock=2.0.0=pypi_0
ncurses=6.1=he6710b0_1
numpy=1.16.3=pypi_0
openssl=1.0.2r=h7b6447c_0
pbr=5.2.0=pypi_0
pip=10.0.1=py35_0
protobuf=3.7.1=pypi_0
pyglet=1.3.2=pypi_0
pyparsing=2.4.0=pypi_0
python=3.5.6=hc3d631a_0
python-dateutil=2.8.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.21.0=pypi_0
scipy=1.2.1=pypi_0
setuptools=40.2.0=py35_0
six=1.12.0=pypi_0
sqlite=3.28.0=h7b6447c_0
tabulate=0.8.3=pypi_0
tensorboard=1.13.1=pypi_0
tensorflow-estimator=1.13.0=pypi_0
tensorflow-gpu=1.13.1=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hbc83047_0
tqdm=4.31.1=pypi_0
urllib3=1.24.2=pypi_0
werkzeug=0.15.2=pypi_0
wheel=0.31.1=py35_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
